+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [CalVer](https://calver.org/about.html) versioning.
 
+## [unreleased]
+
+## Fixed
+
+- `pf2ePlayer.pf2eActor.knownFormulas` threw an error when a formula is non-exitant, probably because it is a custom one.
+
 ## [2025.4.1] - 2025-04-22
 
 ## Changed