+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [CalVer](https://calver.org/about.html) versioning.
 
 - `pf2ePlayer.pf2eActor.knownFormulas` threw an error when a formula is non-exitant, probably because it is a custom one.
 - `Pf2eHelper.pf2eActor._strike` returns a (empty) `sneakAttackDamage` property, which was displayed by `ScribeProvider.scribeCharacterStrike.scribify` (Thank you @master-of-old)
+- `Pf2eHelper.pf2eActor.classDC` failed on retrieving `actor.classDC.attributeModifier.value` when `attributeModifier` is undefined (Thank you @cass-zdanko)
 
 ## [2025.4.1] - 2025-04-22